extracted game over screen
